+++ src/profile/components/ChangeTimezone.js
@@ -2,10 +2,16 @@ import moment from 'moment-timezone';
 import React, { PropTypes, Component } from 'react';
 
 import { Card, CardHeader } from 'linode-components/cards';
-import { Form, FormGroup, Select, SubmitButton } from 'linode-components/forms';
+import {
+  Form,
+  FormGroup,
+  FormSummary,
+  Select,
+  SubmitButton,
+} from 'linode-components/forms';
 
 import { profile } from '~/api';
-import { dispatchOrStoreErrors, FormSummary } from '~/components/forms';
+import { dispatchOrStoreErrors } from '~/api/util';
 import { setStorage } from '~/storage';