+++ src/profile/layouts/LishPage.js
@@ -1,9 +1,14 @@
 import React, { Component } from 'react';
 
 import { Card, CardHeader } from 'linode-components/cards';
-import { Form, FormGroup, FormGroupError, SubmitButton, Select } from 'linode-components/forms';
-
-import { FormSummary } from '~/components/forms';
+import {
+  Form,
+  FormGroup,
+  FormGroupError,
+  FormSummary,
+  SubmitButton,
+  Select,
+} from 'linode-components/forms';
 
 
 export default class NotificationsPage extends Component {